Layered file system (#27)
This PR implements a layered file system, roughly `layered::FileSystem<Upper, Lower>` (closes #5).

Essentially, a layered filesystem itself doesn't carry or store any of the files, but delegates to each of the the layers. Specifically, this implementation will look for and work with files in the upper layer, unless they don't exist, in which case the lower layer is looked at.

The current design of layering treats the lower layer as read-only and performs copy-on-write semantics for moving to upper layer.

The biggest complexity of this implementation comes from needing to support the deletion of files from the lower layer (without actually changing the lower layer), as well as supporting aliasing (same file opened twice) under the presence of writing. There are tests added for ensuring that the behavior of these are as expected.

Currently, the implementation's largest downside is in its handling of directories and permissions, each of which would be improved by having support for `stat` in the core `FileSystem` trait, but I decided that should be a separate PR and not be part of this one. There are relevant TODOs in the code for this. We also would need to handle correct positioning of the "migrated" CoW files, but that would require having support for `lseek` in the core `FileSystem` trait; similarly, future update.