Simplify lifetimes (#91)
This PR gets rid of the `'platform` lifetime we were carrying
everywhere. It was originally set up for full generality, but is
actively annoying and problematic in setting up a few upcoming changes.

In general, we are never really going to want to spin up a platform,
spin it down, and spin up another platform. If that is ever the case, we
might as well start up a full new process.

Thus, this PR eliminates `'platform` from the codebase; wherever
necessary, we set it up as a `'static` lifetime, but generally speaking,
this should just simplify usage non-trivially.

The other change made in this PR is to have a `Box::leak(Box::new(...))`
as part of platform initialization in order to give us this `'static`
lifetime for the linux kernel, linux user, and mock platforms. Given the
requirement of `'static` everywhere now, we must guarantee that the
platforms never disappear, and doing this via a leak is the canonical
way to do it. Interestingly, the Linux userland platform was already
doing this in the runner and such, but now this is guaranteed by
literally the `new` constructors themselves.