Use a shared FD storage within LiteBox's subsystems (#180)
This is part 1 of multiple PRs towards migrating to a shared FD storage
(see #31). I am splitting this into multiple parts in order to not have
the final PR be _too_ massive to review (or indeed even just manage).
However, unfortunately, the overall migration to this is a fairly deep
change, thus it is inevitable that there is going to be non-trivial
amounts to review.

This specific PR focuses on setting up the actual shared storage itself
(i.e., `Descriptors`) and migrating the various `litebox` crate
subsystems into using it. The cleanup for the shims (to remove their
extra storage) will be in a future PR.

Essentially the `Descriptors` storage supports storing descriptor
entries, and managing FDs for any `FdEnabledSubsystem` (a new trait
introduced in this PR). A subsystem can become an `FdEnabledSubsystem`
simply by invoking a macro, simplifying the addition/update of
subsystems, or indeed even if we need to add additional functionality to
the shared FD storage.

Each `FdEnabledSubsystem` can define the type of a descriptor entry, and
the shared descriptor storage is able to handle these varied types
automatically into a shared storage. Additionally, the file descriptors
provided by `Descriptors` storage are all strongly typed (i.e., you
cannot accidentally use a socket FD for a file-system; or even use an
in-mem-fs FD for a device). This strong typing discipline not only
allows us to get strong guarantees from ownership, but also good
ergonomics for getting descriptor entries from within a subsystem. In
particular, despite the shared storage, when accessing with a
strongly-typed FD, when a subsystem asks for an entry, it gets back an
entry of the correct type, and it cannot _accidentally_ access a
different type of descriptor entry than what it is aware of.

This PR also sets up the basics for removing the extra storage that
exists in the Linux shim. This is the "raw integer" interfaces that I've
set up within the `Descriptors`. The actual usage of the raw integer
interfaces will show up in a future PR.

---

### Expected API breakages

- Some of the types no longer implement the `*UnwindSafe` marker traits:
this is not an issue, since we don't expect to `catch_unwind` them
anyways. Mostly just a technicality in terms of breakage.
- We no longer have `FileFd`/`SocketFd` as two types, instead `TypedFd`
is the parametric type to be used. This actually increases our precision
with strong-typing (in the past, you could use a `FileFd` produced by
one file system in another file system; this would have had weird
effects; we now actually statically ensure this cannot happen).
- `WaitableBuilder` no longer needs two interfaces.

These breakages are all accounted for in the shims above them; I am
writing the above notes mostly for documentation, and to note that the
semver-checks _will_ complain about these.