Fix memory manager (#221)
If the memory manger runs in user mode, it is not aware of the memory
used by global allocator and litebox itself (which is allocated by the
host) and thus may choose an unavailable memory address. This PR makes
the host to decide on the address to be allocated.

One tricky problem is how do we make sure the trampoline code can be
loaded at the expected address. Previously, we assumed our memory
manager knows all allocated memory and thus can pick a suitable address
for the binary with extra space after it. One potential solution is to
always `mmap` more pages when mapping a file and later on, use
`MAP_FIXED` when loading the trampoline code. In this PR, I just added
an assertion so that it would panic if loading the trampoline code
fails. The chance that the trampoline code collides with global
allocator's memory should be small as all the trampoline code is loaded
at the beginning.