Implement futex-like subsystem `FutexManager` (#280)
This PR introduces a new `FutexManager` subsystem to LiteBox, which is
designed to make it simpler to provide a futex-like shim north of
LiteBox, without needing to expand the south platform interface.
Importantly, this removes the need for #239 to expand the platform
interface in order to support futexes, by instead providing
library-OS-level futexes.

A natural question one might ask is "why do we need this, when we are
expecting futexes from the platform below?". The actual interface we get
from the layer below is actually narrower than the full Linux futex
interface. Indeed, this is exactly why the (current) status of #239 was
to expand the interface below to get the wider interface needed above.
This would not be sustainable or too maintainable in the long run, since
obviously, not all platforms we support are able to provide the full
futex API. This means that we need to be able to provide a richer futex
API to the north, while getting a narrower futex-like API from the
south; and this means that we need to implement our own version of
futexes.

Another natural question is: should this sit within LiteBox, or should
this sit within the Linux shim. The correct positioning of this is
arguable, but I believe that there is value (especially after some
possible future cleanups) to keep this at the LiteBox layer, since then
our higher-level primitives could depend on this new `FutexManager`
interface---indeed, this might allow us to tighten up our south
interface even more. More design thought is needed, but especially since
that is more of a nice-to-have, I did not want to block things on that
at this point.

In short: this PR provides a richer futex interface to the north,
_without_ changing the south interface for raw mutexes.

Note: there are some optimization opportunities in this implementation
that I have not yet fully squeezed out (there are TODO comments left for
some of these). The reason for this is that some of the details to even
get it working were _very_ subtle, and took a lot of nasty debugging.
Indeed, debugging this is what led to #279, which needs to be merged
before this PR is merged.