Add file truncation support (#318)
This PR implements support for `O_TRUNC` at the shim layer, by
introducing a new `truncate` function at the LiteBox layer. This leads
to a simpler design for truncation at the layered FS, and makes sure
there are no entry leaks upon truncation.

This is similar to #313 but takes the different approach I was mentioning, in order to keep things simpler.