Optimize tar file reading via mmap (#334)
This PR switches the Linux runner to start using mmap to read-in the
initial tar, rather than copying it into memory. This should trivially
help perf especially on large tar files.

To keep things nice, especially since `empty_tar_file` was essentially
just a convenience constant, I've updated it to `EMPTY_TAR_FILE`. We
probably could have updated this when we switched to the `Cow` based
interface for initializing the tar-ro FS, but yeah, this is the better
approach for it.