New `dev_tests` crate + ratcheting for globals (#357)
This PR does a couple of things:
- Moves the ratcheting from the CI yaml stuff to a crate where we can
more easily manage such scripts, allowing more expressivity and nicer
error messages and such. It also makes it such that folks will notice
this locally sooner than waiting on the CI. It is triggered by running
regular tests.
- Fixes #356 by adding the ratchet by looking for any `static` at the
start of the line (so that it doesn't accidentally trigger on `'static`
and such). Note that I explicitly am not looking for `const` which is
also a way to make globals, but such globals are fully read-only, so
don't have the kind of downsides that come from `static` or `static mut`
things.