Caching for pre-processing in runner tests (#363)
The runner tests have some pre-requisites such as compiling C files and
making tar files. Non-trivial chunk of testing time is spent simply
re-running these pre-processing bits of the tests.

This PR adds caching for them (as well as for the rewriter). This, at
least locally for me, speeds up the testing time by ~4x.

Note: non-trivial chunk of this code was written by an LLM. I just
guided it to not make bad decisions. It made a couple of minor bad ones
initially but after a little bit of prompting it did a reasonable
initial draft. The biggest thing it missed even then is that it would
not invalidate the cache for the rewriter if there is a source change in
the rewriter. After prompting it to fix that though, it did the right
thing.