Check for `$ORIGIN` in library paths of Python executable in test runner (#479)
Due to #477 and #478, check for Python versions that the specific issues
and bail out early rather than mark the test as a failure. This test
actually only checks for #478 since it covers at least both things that
I've seen.

To my understanding, the CI's Python version does NOT have this
`$ORIGIN` on it, so the CI will continue testing on Python, but at least
this prevents development blockers locally until #477 and #478 are
fixed.