Track threads per-process and kill them on exit (#518)
Now that we can reliably wake up threads that are waiting or are running
in the guest, let's leverage that capability when terminating a process
via `exit_group`. Track the list of threads in a process, and when a
group exit is requested, mark all threads as exiting and interrupt them.

Also, provide a mechanism for a runner to wait on all threads for a
process to exit and to retrieve the process exit code. Propagate this
exit code to the userland runner's process, and update our tests to
validate a successful exit.

Finally, update `execve` so that it terminates all other threads before
loading the new program.