Close pending sockets (#527)
#497 adds support for graceful close so that `close` becomes blocking
when `SO_LINGER` is enabled and the socket has pending data to be sent.
However, there is no notification after all the data are transferred,
and thus `close` might be blocked indefinitely.

To fix it, we need to check and close those sockets when performing
`check_and_update_events`. Closed sockets would then send out
`Events::HUP` to unblock the caller. To determine whether a socket is
closed or not, we should use `is_open` instead of just checking `state()
== Closed` because `TimeWait` is also considered as closed (which is
handled by `is_open`).